Move block hash types to alloy-eips
